refactor: product listing filters moved to one component
